+++ client/public/manifest.json
