+++ package.json
@@ -9,6 +9,7 @@
 	},
 	"license": "ISC",
 	"dependencies": {
+		"body-parser": "^1.18.3",
 		"cors": "^2.8.5",
 		"express": "^4.16.4",
 		"knex": "^0.15.2",
